refactor: move network logic from customer app to core package
- Relocate Dio client and interceptors (auth, error) to the core package
- Introduce `CoreStorageKeys` for shared storage constants
- Parameterize `refreshTokenPath` in `AuthInterceptor` for improved flexibility
