+++ packages/core/lib/data/storage/core_storage_keys.dart
@@ -1,0 +1,7 @@
+abstract class CoreStorageKeys {
+  CoreStorageKeys._();
+
+  static const String authToken = 'auth_token';
+  static const String refreshToken = 'refresh_token';
+  static const String isLoggedIn = 'is_logged_in';
+}
